% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_manuscript.R
\name{add_manuscript}
\alias{add_manuscript}
\title{Add Rmd Template for PDF Manuscript}
\usage{
add_manuscript(project_path)
}
\arguments{
\item{project_path}{the absolute directory path of where the project was created using `myProject::init()`}
}
\description{
Add a template file for create a rmarkdown pdf manuscript/report
}
