% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_presentation.R
\name{add_presentation}
\alias{add_presentation}
\title{Add Rmd Report}
\usage{
add_presentation(project_path, type = "")
}
\arguments{
\item{project_path}{the absolute directory path of where the project was created using `myProject::init()`}

\item{type}{character}
}
\description{
Add a template file for create a rmarkdown html report
}
