% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_report.R
\name{add_report}
\alias{add_report}
\title{Add Rmd Report}
\usage{
add_report(project_path)
}
\arguments{
\item{project_path}{the absolute directory path of where the project was created using `myProject::init()`}
}
\description{
Add a template file for create a rmarkdown html report
}
