% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create New Project}
\usage{
init(project_path)
}
\arguments{
\item{project_path}{a character vector of the directory where the `drake` & `renv` project will be created}
}
\description{
Initialise a new project
}
\examples{
\dontrun{
init('/home/my_new_project')
}
}
