% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_peakflow_data}
\alias{load_peakflow_data}
\title{Load Peak Flow Data}
\usage{
load_peakflow_data(data = NULL)
}
\arguments{
\item{data}{Optional. A data frame of river flow data. Must contain columns:
"date", "parkerdam", "greenriver", "cameo", "gunnison".}
}
\value{
A data frame with the loaded data.
}
\description{
Load the included peak flow dataset or use user-supplied data.
}
