% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyProject-package.R
\docType{package}
\name{MyProject-package}
\alias{MyProject}
\alias{MyProject-package}
\title{MyProject: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lorenzo-pellis/MyProject}
  \item Report bugs at \url{https://github.com/lorenzo-pellis/MyProject/issues}
}

}
\author{
\strong{Maintainer}: Lorenzo Pellis \email{lorenzo.pellis@manchester.ac.uk} (\href{https://orcid.org/0000-0002-3436-6487}{ORCID}) [copyright holder]

}
\keyword{internal}
