% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_copulas}
\alias{analyze_copulas}
\title{Perform Uniformity Test and Copula Fitting}
\usage{
analyze_copulas(data, params_list)
}
\arguments{
\item{data}{Data frame of flows.}

\item{params_list}{Parameter list from fit_pearson_params.}
}
\value{
A list containing fitted copulas and a summary table.
}
\description{
This function takes uniform data transformed using Pearson distributions and fits various copulas.
}
