% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_pearson_params}
\alias{fit_pearson_params}
\title{Fit Pearson Distributions to Each River}
\usage{
fit_pearson_params(data)
}
\arguments{
\item{data}{Data frame with columns for each river.}
}
\value{
A list of parameter sets for each river.
}
\description{
Fits Pearson distributions to each river's flow data using Maximum Likelihood Estimation.
}
