% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{plot_histograms_with_density}
\alias{plot_histograms_with_density}
\title{Plot Histograms with Density Lines Based on Pearson Fits}
\usage{
plot_histograms_with_density(data, params_list)
}
\arguments{
\item{data}{The data frame.}

\item{params_list}{A list of parameter sets returned by fit_pearson_params.}
}
\value{
A grid of ggplot histograms with density lines.
}
\description{
Plots histograms for each river's flow data and overlays density lines
based on fitted Pearson distributions.
}
