% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_river_histograms}
\alias{plot_river_histograms}
\title{Plot Histograms for Each River}
\usage{
plot_river_histograms(data)
}
\arguments{
\item{data}{A data frame with appropriate columns.}
}
\value{
A grid of ggplot histograms.
}
\description{
Plot histograms for each river column.
}
