% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot Time Series of Peak Flows}
\usage{
plot_time_series(data)
}
\arguments{
\item{data}{A data frame with columns "date", "parkerdam", "greenriver",
"cameo", and "gunnison".}
}
\value{
A ggplot object.
}
\description{
Plot a time series of flow rates for multiple rivers.
}
